pyproject.toml
[project]
name = "blog-webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alt-profanity-check>=1.7.2",
    "django==6.0",
    "django-cleanup>=9.0.0",
    "django-cotton>=2.6.1",
    "django-htmx==1.27.0",
    "django-widget-tweaks>=1.5.1",
    "environs[django]>=14.5.0",
    "pillow>=12.1.0",
    "python-dateutil>=2.9.0.post0",
]
